requirements.txt
litellm
instructor
pydantic
todoist_api_python 
colorama
python-dotenv
